Code Cleaning
Fixed all warnings.
Adjusted DataSource and DataSourceUtils to work a bit cleaner.
DataSources are now reliant on DataSourceUtils having a base to pull data from.
This is good because it will not error when trying to create a DataSource without a base to pull data from.
This is bad because it will not pull data from anything that is not a base.
In the future it will behoove us to loosen that coupling .